requirements.txt
reportlab
mysqlclient
xhtml2pdf
html5lib
pillow
beautifulsoup4
django-summernote
easy-thumbnails
